Add a colour chart to the page
Note: this only charts the first sample in the uploaded CSV

Getting the chart to have a background fill colour that scaled when the page
was resized proved difficult.

I found some code on codepen[1] that pointed me in the right direction. We
patch the draw method for Chart.js's line element to create a colour gradient
with the correct dimensions for the chart and then set that as the backgroundColour

A better solution is likely needed since this solution only works with chart.js
2.5.0 which is a few years old.

[1] https://codepen.io/jordanwillis/pen/BWxErp